spatie permission install roles & seeders middlewares routes

+++ routes/api.php
@@ -18,17 +18,17 @@ Route::post('login', [UsserController::class, 'login']);
 
 Route::middleware('auth:api')->group(function() {
    
-    Route::put('/players/{id}',[UserController::class, 'update']); //modifica nom jugador
-    Route::post('/players/{id}/games',[UserController::class, 'play']); //jugador realitza una jugada 
-    Route::delete('/players/{id}/games',[UserController::class, 'destroy']); //elimina totes tirades jugador X
-    Route::get('players',[UserController::class, 'index']); //retorna tots jugadors
-    Route::get('/players/{id}/games', [UserController::class, 'index']); // llistat jugades jugador X
-    Route::get('/plaers/ranking',[UserController::class, 'ranking']); // % mitjà d'exits
-    Route::get('/plaers/ranking/loser',[UserController::class, 'ranking.loser']); // jugador amb % exit mes baix
-    Route::get('/plaers/ranking/winner',[UserController::class, 'winner']); //jugador amb % exit mes alt
+    Route::put('/players/{id}',[UserController::class, 'update']) ->middleware('role:player'); //modifica nom jugador
+    Route::post('/players/{id}/games',[UserController::class, 'play'])->middleware('role:player'); //jugador realitza una jugada 
+    Route::delete('/players/{id}/games',[UserController::class, 'destroy'])->middleware('role:player'); //elimina totes tirades jugador X
+    Route::get('players',[UserController::class, 'index'])->middleware('role:admin'); //retorna tots jugadors
+    Route::get('/players/{id}/games', [UserController::class, 'index'])->middleware('role:player'); // llistat jugades jugador X
+    Route::get('/plaers/ranking',[UserController::class, 'ranking'])->middleware('role:admin'); // % mitjà d'exits
+    Route::get('/plaers/ranking/loser',[UserController::class, 'ranking.loser'])->middleware('role:admin'); // jugador amb % exit mes baix
+    Route::get('/plaers/ranking/winner',[UserController::class, 'winner'])->middleware('role:admin'); //jugador amb % exit mes alt
 
 
 
-    Route::post('logout',[UserController::class, 'logout']);
+    Route::post('logout',[UserController::class, 'logout'])->middleware('role:admin|player');
 });
